添加角色

+++ ep-api/src/main/java/com/all580/role/api/service/FuncService.java
@@ -24,6 +24,6 @@ public interface FuncService {
 
     Result updateByPrimaryKeySelective(Map<String,Object> params);
 
-    Result deleteByPrimaryKey(Long id);
+    Result deleteByPrimaryKey(int id);
 
 }
